requirement.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.2
Django==5.2.4
djangorestframework==3.16.0
idna==3.10
iniconfig==2.1.0
mysql-connector-python==9.4.0
mysqlclient==2.2.7
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
PyMySQL==1.1.1
pytest==8.4.1
requests==2.32.4
sqlparse==0.5.3
urllib3==2.5.0
